requirements.txt
requests>=2.31.0
langchain>=0.1.0
langchain-openai>=0.0.2
openai>=1.0.0
alpaca-trade-api>=3.0.0
python-dotenv>=1.0.0